requirements.txt
boto3>=1.26.146
docopt>=0.6.2
dohq_artifactory>=0.8.4
kubernetes>=26.1.0
pandas>=2.0.2
xlsxwriter>=3.1.2